Edit counter template making plus symbols optional
In the original template, each of the counter items are appended with a
a plus symbol. Since we include post-docs and interns in the counter,
the plus symbol should be optional, since it is clumsy to refer to the
number of SIParCS interns as 3+. DAReS should have an exact count of
personnel. Include the plus symbols in the data. Make the background
color the primary color of the site design.